# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Swagger Doc For Find-Apps Api ",
        "version": "1.0.0",
        "title": "Restaurant Apis"
    },
...
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "paths": {
...
        "/user/verifyOtp": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": " verifyOtp",
                "description": "verifyOtp ",
                "operationId": "verifyOtp",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/user"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/user"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
...
        "/user/resendOtp": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": " updateUser",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "updateUser ",
                "operationId": "updateUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/user"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/user"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
        "/user/updateUser": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": " updateUser",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "updateUser ",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/user"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/user"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
...
    },
...
    "servers": [
        {
            "url": "http://localhost:4000"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "in": "header",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "user": {
                "type": "object",
                "required": [
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "1"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "9996663330"
                    },
                    "name": {
                        "type": "string",
                        "example": "Harsh"
                    },
                    "age": {
                        "type": "number",
                        "example": "23"
                    },
                    "email": {
                        "type": "string",
                        "example": "example12@gmail.com"
                    },
                    "address": {
                        "type": "string",
                        "example": "hno 147 mohali tower near shoppig mall punjab"
                    },
                    "gender": {
                        "type": "string",
                        "example": "male"
                    },
                    "otp": {
                        "type": "string",
                        "example": "789456"
                    }
                }
            },
...
        }
    }
}